requirements.txt
django == 2.2.7
djangorestframework == 3.10.3
factory == 2.12.0